+++ examples/helpers/createMesh.ts
@@ -1,6 +1,6 @@
 import { SpiderMesh, Topology } from '@spider-mesh/core'
 import { EventBus } from '@spider-mesh/events'
-import { TopologyDiscoveryAdapter } from '@spider-mesh/discovery'
+import { TopologyDiscoveryAdapter } from '../../src/index.js'
 import { Http2Pubsub, Http2Rpc } from '../../src/index.js'
 import { createDiscovery } from './createDiscovery.js'